--- pages/components/logic/logic
@@ -1,2 +1,0 @@
-
-const
