--- pages/components/test/test1.tsx
@@ -1,210 +1,0 @@
-import React from "react"
-
-const test1 = {
-    "test1": [{     
-        "position": 1, 
-        "question": "Para mí es importante recibir la aprobación de los demás.",
-        "interpretation": {"BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO":"Para un adulto es absolutamente necesario tener el cariño y la aprobación de sus semejantes, familia y amigos. "},
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 2, 
-        "question": "Odio equivocarme en algo.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO", 
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 3, 
-        "question": "La gente que se equivoca, logra lo que se merece.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{ 
-        "position": 4, 
-        "question": "Generalmente acepto los acontecimientos con filosofía.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{ 
-        "position": 5, 
-        "question": "Si una persona quiere, puede ser feliz en casi cualquier circunstancia.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 6, 
-        "question": "Temo a las cosas que, a menudo, me resultan objeto de preocupación.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 7, 
-        "question": "Normalmente aplazo las decisiones importantes.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 8, 
-        "question": "Todo el mundo necesita de alguien a quién recurrir en busca de ayuda y consejo.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 9, 
-        "question": "“Una cebra no puede cambiar sus rayas”.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 10, 
-        "question": "Prefiero sobre todas las cosas pasar el tiempo libre de una forma tranquila.",
-        "interpretation": "BUSQUEDA DE APROBACIÓN Y RECONOCIMIENTO", 
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 11, 
-        "question": "Me gusta que los demás me respeten, pero yo no tengo porque respetar a nadie.",
-        "interpretation": {"PERFECCIONISTA":"Debe ser indefectiblemente competente y casi perfecto en todo lo que emprende."},
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 12, 
-        "question": "Evito las cosas que no puedo hacer bien.",
-        "interpretation": "PERFECCIONISTA", 
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 13, 
-        "question": "Hay demasiadas personas malas que escapan del castigo del infierno.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 14, 
-        "question": "Las frustraciones no me distorsionan.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 15, 
-        "question": "A la gente no le trastornan los acontecimientos sino la imagen que tiene de ellos.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 16, 
-        "question": "Me producen poca ansiedad los peligros inesperados o los acontecimientos futuros.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 17, 
-        "question": "Trato de afrontar los trabajos fastidiosos y hacerlos cuanto antes.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 18, 
-        "question": "En las decisiones importantes, consulto con una autoridad al respecto.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 19, 
-        "question": "Es casi imposible superar la influencia del pasado.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 20, 
-        "question": "Me gusta disponer de muchos recursos.",
-        "interpretation": "PERFECCIONISTA",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 21, 
-        "question": "Quiero gustar a todo el mundo.",
-        "interpretation": {"CONDENACIÓN":"Ciertas personas son malas, viles y perversas y deberían ser castigadas. "},
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 22, 
-        "question": "No me gusta competir en actividades en las que los demás son mejores que yo.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 23, 
-        "question": "Aquellos que se equivocan merecen cargar con la culpa.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 24, 
-        "question": "Las cosas deberían ser distintas a como son.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 25, 
-        "question": "Yo provoco mi propio mal humor.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 26, 
-        "question": "A menudo, no puedo quitarme algún asunto de la cabeza.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 27, 
-        "question": "Evito enfrentarme a los problemas.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 28, 
-        "question": "Todo el mundo necesita tener fuera de si mismo una fuente de energía.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 29, 
-        "question": "Solo porque una vez algo afecto tu vida de forma importante, no quiere decir que tenga que ser igual en el futuro.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 30, 
-        "question": "Me siento mas satisfecho cuando tengo muchas cosas que hacer.",
-        "interpretation": "CONDENACIÓN",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 31, 
-        "question": "Puedo gustarme a mi mismo aun cuando no guste a los demás.",
-        "interpretation": {"BTF": "Lo que mas me divierte es realizar algún proyecto creativo."},
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 32, 
-        "question": "Me gustaría triunfar en algo, pero no pienso que deba hacerlo.",
-        "interpretation": "BTF",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 33, 
-        "question": "La inmoralidad debería castigarse severamente.",
-        "interpretation": "BTF",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 34, 
-        "question": "A menudo me siento trastornado por situaciones que no me gustan.",
-        "interpretation": "BTF",
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 35, 
-        "question": "Las personas desgraciadas, normalmente, se deben este estado a si mismas.",
-        "interpretation": "BTF",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 36, 
-        "question": "No me preocupo por no poder evitar que algo ocurra.",
-        "interpretation": "BTF",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 37, 
-        "question": "Normalmente tomo las decisiones tan pronto como puedo.",
-        "interpretation": "BTF",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 38, 
-        "question": "ay determinadas personas de las que dependo mucho.",
-        "interpretation": {"BTF": "Lo que mas me divierte es realizar algún proyecto creativo."},
-        "answerstype": {"A": [{"No": 0}, {"Yes": 1}]}
-    },{
-        "position": 39, 
-        "question": "La gente sobrevalora la influencia del pasado.",
-        "interpretation": "BTF",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    },{
-        "position": 40, 
-        "question": "Lo que mas me divierte es realizar algún proyecto creativo.",
-        "interpretation": "BTF",
-        "answerstype": {"B": [{"No": 1}, {"Yes": 0}]}
-    }]
-}
-
-
-
-    
-export default test1
