--- pages/home/index.tsx
@@ -1,8 +1,0 @@
-
-const Home = () => {
-    return (
-        <h1> Este es la pagina de home </h1>
-    )
-}
-
-export default Home
