+++ pages/index.tsx
@@ -1,71 +1,18 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
-
-export default function Home() {
+const index = () => {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Brain Test</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
+      <>
+          <h1>TEST DE CREENCIAS</h1>
+          <h2> INSTRUCCIONES PARA RESPONDER AL CUESTIONARIO </h2>
+          <ul>
+              <li>•	No hay límite de tiempo para contestar al Cuestionario.</li>
+              <li>•	No hay respuestas correctas o erróneas. Será útil en la medida que sea sincero/a en sus respuestas.</li>
+              <li>•	No es necesario que piense mucho rato en cada ítem.</li>
+              <li>•	Si está más de acuerdo que en desacuerdo con el ítem señale la casilla SI, si, por el contrario, está más en desacuerdo que de acuerdo, señale la casilla NO.</li>
+              <li>•	Asegúrese de que contesta lo que usted realmente piensa, no lo que cree que debería pensar.</li>
+              <li>•	Conteste a todos los ítems.</li>
+          </ul>
+          <button>proceder con el test</button>   
+      </>
   )
 }
+export default index
