+++ pages/test/index.tsx
@@ -1,0 +1,8 @@
+import TestLayaout from "../../src/components/layaouts/TestLayaout"
+
+const Test = () => {
+
+    return(<TestLayaout />)
+}
+
+export default Test
